+++ src/GameCore/GAMECORE_CST.ts
@@ -1,0 +1,6 @@
+export namespace GAMECORE_CST {
+	export const enum Norm {
+		Euclidian = "Euclidian",
+		Taxicab = "Taxicab",
+	}
+}
