+++ src/GameCore/Race/ClientRaceGameController.ts
@@ -1,6 +1,6 @@
-import Player from "../../GameCore/player";
-import RaceGameController from "../../GameCore/Race/RaceGameController";
-import RaceGameState from "../../GameCore/Race/RaceGameState";
+import Player from "./player";
+import RaceGameController from "./RaceGameController";
+import RaceGameState from "./RaceGameState";
 
 export default class ClientRaceGameController extends RaceGameController {
 	constructor(players: Player[] = []) {
