+++ src/GameCore/Race/RACE_CST.ts
@@ -1,5 +1,8 @@
+import { GAMECORE_CST } from "../GAMECORE_CST";
+
 export const RACE_CST = {
 	MOVE: {
-		SPEED: 10, //in tile per millisecond.
+		SPEED: 0.001, //in tile per millisecond.
+		NORM: GAMECORE_CST.Norm.Taxicab,
 	},
 };
