+++ src/GameCore/Race/point.ts
